:recycle: separate move logic
