+++ src/components/canvas/Metaball/before/Metaballs.ts
@@ -1,4 +1,5 @@
 import {DynamicMetaball, StaticMetaball} from './Metaball';
+import {MoveStrategy} from './Strategies';
 
 abstract class Metaballs<MetaballType> {
   abstract balls: MetaballType[];
@@ -6,43 +7,6 @@ abstract class Metaballs<MetaballType> {
   abstract push(metaball: MetaballType): void;
 
   abstract moveAll(): void;
-}
-
-abstract class Strategy {
-  abstract exec(...args: unknown[]): void;
-
-  abstract before?: (...args: unknown[]) => void;
-
-  abstract after?: (...args: unknown[]) => void;
-}
-
-export class MoveStrategy implements Strategy {
-  before?: (...args: unknown[]) => void;
-
-  after?: (...args: unknown[]) => void;
-
-  constructor() {}
-
-  setBefore(callback: (...args: unknown[]) => void) {
-    this.before = callback.bind(this);
-  }
-
-  setAfter(callback: (...args: unknown[]) => void) {
-    this.after = callback.bind(this);
-  }
-
-  // before() {}
-
-  // after() {}
-
-  exec(metaball: DynamicMetaball) {
-    // console.log('this: ', metaball, this);
-    this.before?.();
-
-    metaball.move();
-
-    this.after?.();
-  }
 }
 
 export class StaticMetaballs implements Metaballs<StaticMetaball> {
